requirements.txt
flask==3.0.0
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.4
google-generativeai==0.3.2
python-dotenv==1.0.0
numpy==1.25.2
scipy==1.11.4
duckdb==0.9.2
openpyxl==3.1.2
python-multipart==0.0.6
werkzeug==3.0.1
gunicorn==21.2.0
flask>=2.3.0
werkzeug>=2.3.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Web scraping
requests>=2.28.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# AI integration
google-generativeai>=0.3.0

# Utilities
python-dotenv>=1.0.0
openpyxl>=3.1.0
python-multipart>=0.0.6

# Database (optional - can be removed if causing issues)
# duckdb>=0.9.0

# Production server
gunicorn>=21.0.0